requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bedrock-agentcore==0.1.1
bedrock-agentcore-starter-toolkit==0.1.2
boto3==1.39.13
botocore==1.39.13
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.2
mdurl==0.1.2
opentelemetry-api==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-threading==0.56b0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
s3transfer==0.13.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.4.1
starlette==0.47.2
strands-agents==1.1.0
toml==0.10.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.23.0
strands-agents-tools